RenderBlockingAssetSpanDetector
