+++ src/sentry/utils/performance_issues/detectors/__init__.py
@@ -1,5 +1,6 @@
 from .consecutive_db_detector import ConsecutiveDBSpanDetector  # NOQA
 from .consecutive_http_detector import ConsecutiveHTTPSpanDetector  # NOQA
 from .n_plus_one_api_calls_detector import NPlusOneAPICallsDetector  # NOQA
+from .render_blocking_asset_span_detector import RenderBlockingAssetSpanDetector  # NOQA
 from .slow_db_query_detector import SlowDBQueryDetector  # NOQA
 from .uncompressed_asset_detector import UncompressedAssetSpanDetector  # NOQA
