+++ tests/sentry/utils/performance_issues/test_render_blocking_asset_detector.py
@@ -11,9 +11,9 @@ from sentry.testutils.performance_issues.event_generators import (
     modify_span_start,
 )
 from sentry.testutils.silo import region_silo_test
+from sentry.utils.performance_issues.detectors import RenderBlockingAssetSpanDetector
 from sentry.utils.performance_issues.performance_detection import (
     PerformanceProblem,
-    RenderBlockingAssetSpanDetector,
     get_detection_settings,
     run_detector_on_data,
 )
